Load oncoEnrichR annotation database

Load oncoEnrichR annotation database

Usage

init_report(
  oeDB,
  project_title = "_Project title_",
  project_owner = "_Project owner_",
  html_floating_toc = T,
  html_report_theme = "default",
  query_id_type = "symbol",
  ignore_id_err = TRUE,
  project_description = "_Project description_",
  ppi_min_string_score = 900,
  ppi_add_nodes = 50,
  ppi_node_shadow = TRUE,
  bgset_description = "All annotated protein-coding genes",
  bgset_id_type = "symbol",
  p_value_cutoff_enrichment = 0.05,
  p_value_adjustment_method = "BH",
  q_value_cutoff_enrichment = 0.2,
  min_geneset_size = 10,
  max_geneset_size = 500,
  num_terms_enrichment_plot = 20,
  min_subcellcomp_confidence = 1,
  subcellcomp_show_cytosol = F,
  min_confidence_reg_interaction = "D",
  max_fitness_score = -2,
  simplify_go = F,
  show_ppi = T,
  show_drugs_in_ppi = T,
  show_disease = T,
  show_top_diseases_only = T,
  show_cancer_hallmarks = T,
  show_drug = T,
  show_enrichment = T,
  show_aberration = T,
  show_coexpression = T,
  show_subcell_comp = T,
  show_fitness = T,
  show_cell_tissue = F,
  show_ligand_receptor = T,
  show_regulatory = T,
  show_unknown_function = T,
  show_prognostic = T,
  show_synleth = T,
  show_complex = T,
  show_domain = T
)

Arguments

oeDB

oncoEnrichR annotation database object

project_title

project title (title of report)

project_owner

name of project owner

html_floating_toc

logical - float the table of contents to the left of the main document content. The floating table of contents will always be visible even when the document is scrolled

html_report_theme

Bootswatch theme for HTML report (any of "bootstrap","cerulean","cosmo","default","flatly","journal","lumen","paper","sandstone","simplex","spacelab","united","yeti")

query_id_type

character indicating source of query (one of "uniprot_acc", "symbol", "entrezgene", or "ensembl_gene","ensembl_mrna","refseq_mrna","ensembl_protein","refseq_protein")

ignore_id_err

logical indicating if analysis should continue when uknown query identifiers are encountered

project_description

project background information

ppi_min_string_score

minimum score (0-1000) for retrieval of protein-protein interactions (STRING)

ppi_add_nodes

number of nodes to add to target set when computing the protein-protein interaction network (STRING)

ppi_node_shadow

show shadow for nodes in the displayed PPI network

bgset_description

character indicating type of background (e.g. "All lipid-binding proteins (n = 200)")

bgset_id_type

character indicating source of query (one of "uniprot_acc", "symbol", "entrezgene", or "ensembl_gene","ensembl_mrna","refseq_mrna","ensembl_protein","refseq_protein")

p_value_cutoff_enrichment

cutoff p-value for enrichment/over-representation analysis

p_value_adjustment_method

one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

q_value_cutoff_enrichment

cutoff q-value for enrichment analysis

min_geneset_size

minimal size of geneset annotated by term for testing in enrichment/over-representation analysis

max_geneset_size

maximal size of geneset annotated by term for testing in enrichment/over-representation analysis

num_terms_enrichment_plot

number of top enriched Gene Ontology terms (max) to show in enrichment barplot

min_subcellcomp_confidence

minimum confidence level of subcellular compartment annotations (range from 1 to 6, 6 is strongest)

subcellcomp_show_cytosol

logical indicating if subcellular heatmap should highlight cytosol as a subcellular protein location or not

min_confidence_reg_interaction

minimum confidence level for regulatory interactions (TF-target) retrieved from DoRothEA ('A','B','C', or 'D')

max_fitness_score

maximum loss-of-fitness score (scaled Bayes factor from BAGEL) for genes retrieved from Project Score

simplify_go

remove highly similar GO terms in results from GO enrichment/over-representation analysis

show_ppi

logical indicating if report should contain protein-protein interaction data (STRING)

show_drugs_in_ppi

logical indicating if targeted drugs (> phase 3) should be displayed in protein-protein interaction network (Open Targets Platform)

show_disease

logical indicating if report should contain disease associations (Open Targets Platform, association_score >= 0.05, support from at least two data types)

show_top_diseases_only

logical indicating if report should contain top (n = 20) disease associations only pr. query gene (Open Targets Platform)

show_cancer_hallmarks

logical indicating if report should contain annotations/evidence of cancer hallmarks per query gene (COSMIC/Open Targets Platform)

show_drug

logical indicating if report should contain targeted cancer drug information

show_enrichment

logical indicating if report should contain functional enrichment/over-representation analysis (MSigDB, GO, KEGG, REACTOME, NetPath, WikiPathways)

show_aberration

logical indicating if report should contain TCGA aberration plots (amplifications/deletions)

show_coexpression

logical indicating if report should contain TCGA co-expression data (RNAseq) of query set with oncogenes/tumor suppressor genes

show_subcell_comp

logical indicating if report should provide subcellular compartment annotations (ComPPI)

show_fitness

logical indicating if report should provide fitness scores and target priority scores from CRISPR/Cas9 loss-of-fitness screens (Project Score)

show_cell_tissue

logical indicating if report should contain tissue-specificity and single cell-type specificity assessments (Human Protein Atlas) of target genes, using data from the Human Protein Atlas

show_ligand_receptor

logical indicating if report should contain ligand-receptor interactions (CellChatDB)

show_regulatory

logical indicating if report should contain data on transcription factor (TF) - target interactions relevant for the query set (DoRothEA)

show_unknown_function

logical indicating if report should highlight target genes with unknown or poorly defined functions (GO/Uniprot KB/NCBI)

show_prognostic

logical indicating if mRNA-based (single-gene) prognostic associations to cancer types should be listed (Human Protein Atlas/TCGA)

show_synleth

logical indicating if report should list overlap with predicted synthetic lethality interactions (gene paralogs only, De Kegel et al., Cell Systems, 2021)

show_complex

logical indicating if report should provide target memberships in known protein complexes (ComplexPortal/Compleat/PDB/CORUM)

show_domain

logical indicating if report should provide target memberships in known protein domains (Pfam)

remote

If TRUE (default), load database from version-tagged Zenodo repository

cache_dir

path to local cache for more efficient subsequent loads