Write oncoEnrichR report object to output file

Function that writes an oncoEnrichR report object to file, either as an interactive HTML report or as an Excel workbook.

Usage

write(
  report,
  oeDB,
  file = "testReport.html",
  ignore_file_extension = F,
  overwrite = F,
  format = "html",
  ...
)

Arguments

report

object with oncoEnrichR report data (returned by oeDB$onco_enrich)

oeDB

oncoEnrichR data repository object - as returned from load_db()

file

full filename for report output (e.g. "oe_report.html" or "oe_report.xlsx")

ignore_file_extension

logical to accept any type of filaname extensions (for Galaxy integration)

overwrite

logical indicating if existing output files may be overwritten

format

file format of output (html/excel)

...

options for Galaxy/non self-contained HTML. Only applicable for use in Galaxy